requirements.txt
django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary>=2.8
django-environ==0.11.2
drf-yasg==1.21.8
gunicorn==23.0.0
django-cors-headers==4.6.0
django-filter==24.1
pytest
pytest-django
channels
daphne